Move helper functions out of the kopia.go file
